Code cleanup: separate source files for llm functions.
